Support Python 3.6+
