+++ src/outcome/_version.py
@@ -1,4 +1,3 @@
-# coding: utf-8
 # This file is imported from __init__.py and exec'd from setup.py
 
 __version__ = "1.0.1"
