+++ test-requirements.txt
@@ -1,4 +1,3 @@
 pytest
 pytest-cov
-pytest-asyncio; python_version >= '3.5'
-async-generator; python_version >= '3.5'
+pytest-asyncio
