--- tests/__init__.py
@@ -1,2 +1,0 @@
-# coding: utf-8
-from __future__ import absolute_import, division, print_function
